Added RBDC/VBDC attribute checking to SatLowerLayerService. Changed the quantization rules of RBDC and VBDC. Small bug fixes.

+++ model/satellite-ut-scheduler.cc
@@ -124,7 +124,7 @@ SatUtScheduler::DoScheduling (std::vector<Ptr<Packet> > &packets, uint32_t paylo
 {
   NS_LOG_FUNCTION (this << payloadBytes << rcIndex << policy);
 
-  NS_LOG_LOGIC ("UT scheduling RC: " << rcIndex << " with " << payloadBytes << " bytes");
+  NS_LOG_LOGIC ("UT scheduling RC: " << (uint32_t)(rcIndex) << " with " << payloadBytes << " bytes");
 
   // Schedule the requested RC index
   DoSchedulingForRcIndex (packets, payloadBytes, rcIndex);
@@ -142,7 +142,7 @@ SatUtScheduler::DoScheduling (std::vector<Ptr<Packet> > &packets, uint32_t paylo
           // No use asking the given RC index again
           if (*it != rcIndex)
             {
-              NS_LOG_LOGIC ("UT scheduling RC: " << rcIndex << " with " << payloadBytes << " bytes");
+              NS_LOG_LOGIC ("UT scheduling RC: " << (uint32_t)(rcIndex) << " with " << payloadBytes << " bytes");
 
               // Schedule the requested RC index
               uint32_t bytes = DoSchedulingForRcIndex (packets, payloadBytes, *it);
@@ -180,7 +180,7 @@ SatUtScheduler::DoSchedulingForRcIndex (std::vector<Ptr<Packet> > &packets, uint
       Ptr<Packet> p = m_txOpportunityCallback (payloadBytes, m_nodeInfo->GetMacAddress (), rcIndex);
       if (p)
         {
-          NS_LOG_LOGIC ("Created a packet from RC: " << rcIndex << " size: " << p->GetSize ());
+          NS_LOG_LOGIC ("Created a packet from RC: " << (uint32_t)(rcIndex) << " size: " << p->GetSize ());
 
           packets.push_back (p);
